Rename plugin template to dk-market
