+++ app/controllers/dk_market/examples_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-module ::MyPluginModule
+module ::DkMarket
   class ExamplesController < ::ApplicationController
     requires_plugin PLUGIN_NAME
