+++ config/routes.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-MyPluginModule::Engine.routes.draw do
+DkMarket::Engine.routes.draw do
   get "/examples" => "examples#index"
   # define routes here
 end
 
-Discourse::Application.routes.draw { mount ::MyPluginModule::Engine, at: "my-plugin" }
+Discourse::Application.routes.draw { mount ::DkMarket::Engine, at: "dk-market" }
