+++ plugin.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-# name: discourse-plugin-name
+# name: dk-market
 # about: TODO
 # meta_topic_id: TODO
 # version: 0.0.1
@@ -8,13 +8,13 @@
 # url: TODO
 # required_version: 2.7.0
 
-enabled_site_setting :plugin_name_enabled
+enabled_site_setting :dk_market_enabled
 
-module ::MyPluginModule
-  PLUGIN_NAME = "discourse-plugin-name"
+module ::DkMarket
+  PLUGIN_NAME = "dk-market"
 end
 
-require_relative "lib/my_plugin_module/engine"
+require_relative "lib/dk_market/engine"
 
 after_initialize do
   # Code which should run after Rails has finished booting
